+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 - Fix beneficiary list navbar
 - Fix workshops date input
+- Fix workshops hour input
 - Fix Workshops navbar
 - Fix beneficiary registration navBar appearing